Create Pin unit tests

+++ backend/src/main/java/com/luiz/backend/services/PinServiceImpl.java
@@ -11,6 +11,7 @@ import com.luiz.backend.dtos.PinPostRequest;
 import com.luiz.backend.entity.Pin;
 import com.luiz.backend.entity.User;
 import com.luiz.backend.exception.PinNotFoundException;
+import com.luiz.backend.exception.UnauthorizedException;
 import com.luiz.backend.exception.UserNotFoundException;
 import com.luiz.backend.mappers.PinMapper;
 import com.luiz.backend.repository.PinRepository;
@@ -28,6 +29,10 @@ public class PinServiceImpl implements PinService {
 
   @Override
   public PinDto createPin(PinPostRequest request, User user) {
+    if (request == null || user == null) {
+      throw new IllegalArgumentException("Request or user cannot be null");
+    }
+    
     Pin pin = mapper.toEntity(request);
     pin.setUser(user);
 
@@ -45,19 +50,27 @@ public class PinServiceImpl implements PinService {
 
   @Override
   public void deletePin(UUID id, User user) {
+    // TODO Use a custom query to optimize data fetch
     Pin pin = repository.findById(id).orElseThrow(() -> new PinNotFoundException("Pin not found with id " + id));
+
+    if (!pin.getUser().getId().equals(user.getId())) {
+      throw new UnauthorizedException(
+          "User with ID " + user.getId() + " is not authorized to delete pin with ID " + id);
+    }
+
     repository.delete(pin);
   }
 
   @Override
-  public Page<PinDto> getPinsByTag(String tags, Pageable pageable) {
-    Page<Pin> pinsPage = repository.findByTags(tags, pageable);
+  public Page<PinDto> getPinsByTag(String tag, Pageable pageable) {
+    Page<Pin> pinsPage = repository.findByTag(tag, pageable);
     return pinsPage.map(mapper::toDto);
   }
 
   @Override
   public Page<PinDto> getPinsByUser(UUID userId, Pageable pageable) {
-    User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException("User not found with id " + userId));
+    User user = userRepository.findById(userId)
+        .orElseThrow(() -> new UserNotFoundException("User not found with id " + userId));
     Page<Pin> pinsPage = repository.findByUser(user, pageable);
     return pinsPage.map(mapper::toDto);
   }
